refactor: update Carousel using hooks
